refactor: move to platform folders
