+++ lib/config/presets/github/index.ts
@@ -3,7 +3,7 @@ import { PLATFORM_TYPE_GITHUB } from '../../../constants/platforms';
 import { logger } from '../../../logger';
 import { Http, HttpOptions } from '../../../util/http';
 import { ensureTrailingSlash } from '../../../util/url';
-import { Preset } from '../common';
+import { Preset, PresetConfig } from '../common';
 
 const http = new Http(PLATFORM_TYPE_GITHUB);
 
@@ -73,4 +73,11 @@ export async function getPresetFromEndpoint(
     return jsonContent[presetName];
   }
   return jsonContent;
+}
+
+export function getPreset({
+  packageName: pkgName,
+  presetName = 'default',
+}: PresetConfig): Promise<Preset> {
+  return getPresetFromEndpoint(pkgName, presetName);
 }
