--- lib/config/presets/gitlab.ts
@@ -1,9 +1,0 @@
-import { Preset, PresetConfig } from './common';
-import { getPresetFromEndpoint } from './local/gitlab';
-
-export function getPreset({
-  packageName: pkgName,
-  presetName = 'default',
-}: PresetConfig): Promise<Preset> {
-  return getPresetFromEndpoint(pkgName, presetName);
-}
